BrowsingDataRemover::DataType updated to uint64_t
Chromium change:
https://chromium.googlesource.com/chromium/src/+/1706d8e3c831f5e9e309a02a0b2d602deea53c86

commit 1706d8e3c831f5e9e309a02a0b2d602deea53c86
Author: cammie <cammie@chromium.org>
Date:   Mon Apr 25 23:28:59 2022 +0000

    Shared Storage: Integrate with BrowsingDataRemover.

    We hook up with `BrowsingDataRemoverImpl` so that Shared Storage, as
    well as other Privacy Sandbox API persistent storage, will be cleared
    when the clear-site-data header is sent for storage.

    Also, we must update to using uint64_t instead of int in a couple of
    places because we add the 21st value (1 << 20) to
    `BrowsingDataRemover::DataType`.

    Moreover, we add a new
    `BrowsingDataRemover::DataType::DATA_TYPE_PRIVACY_SANDBOX` collection
    and hook this up to be cleared by the Clear-Site-Data header so that
    origin-associated storage from Privacy Sandbox APIs (including
    Shared Storage) will be cleared when "storage" is sent as a value.

    Bug: 1218540

+++ browser/browsing_data/brave_clear_browsing_data_browsertest.cc
@@ -160,7 +160,7 @@ class BraveClearDataOnExitTest
                             true);
   }
 
-  int GetRemoveMaskAll() {
+  uint64_t GetRemoveMaskAll() {
     return chrome_browsing_data_remover::DATA_TYPE_HISTORY |
            content::BrowsingDataRemover::DATA_TYPE_DOWNLOADS |
            content::BrowsingDataRemover::DATA_TYPE_CACHE |
@@ -177,8 +177,8 @@ class BraveClearDataOnExitTest
 
   // BraveClearBrowsingData::OnExitTestingCallback implementation.
   void BeforeClearOnExitRemoveData(content::BrowsingDataRemover* remover,
-                                   int remove_mask,
-                                   int origin_mask) override {
+                                   uint64_t remove_mask,
+                                   uint64_t origin_mask) override {
     remove_data_call_count_++;
 
     if (expected_remove_mask_ != -1)
